Spinner: Move spinner into widgets folder
