+++ tests/unit/spinner/options.js
@@ -1,6 +1,6 @@
 define( [
 	"jquery",
-	"ui/spinner",
+	"ui/widgets/spinner",
 	"globalize",
 	"globalize/ja-JP"
 ], function( $ ) {
